always send from our CLIENT_PORT => relay SERVER_PORT
dhcp relays are allowed by the RFC to expect that all requests
headed to them originate from CLIENT_PORT (and to reply only to
that port). so we should use it for sending.

also we should make sure to always reply from SERVER_PORT when
sending back to clients, since they are also permitted to expect
that.